:recycle: gamble 확률 검증 테스트 분리
